+++ src/Phan/Language/Type/FloatType.php
@@ -61,8 +61,7 @@ class FloatType extends ScalarType
         if ($other instanceof ScalarType) {
             return $other instanceof FloatType || (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
         }
-        return $other instanceof CallableType ||
-            $other instanceof TemplateType ||
+        return $other instanceof TemplateType ||
             $other instanceof MixedType;
     }
 }